move fields util into utils
