+++ src/example-types/results.js
@@ -1,7 +1,7 @@
 let F = require('futil')
 let _ = require('lodash/fp')
 let { highlightResults, arrayToHighlightsFieldMap } = require('../utils/highlighting')
-let { getField } = require('../fields')
+let { getField } = require('../utils/fields')
 
 module.exports = {
   validContext: () => true,
